app/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
clickclick==20.10.2
exceptiongroup==1.1.3
fastapi==0.103.2
flask-restplus==0.13.0
idna==3.4
inflection==0.5.1
injector==0.21.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
swagger-ui==0.1.2
swagger-ui-bundle==0.0.9
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==2.2.3
